requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url requirements.in
#
aioboto3==13.3.0
    # via -r requirements.in
aiobotocore[boto3]==2.16.0
    # via aioboto3
aiodns==4.0.0
    # via aiohttp
aiofiles==25.1.0
    # via aioboto3
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp[speedups]==3.13.3
    # via
    #   -r requirements.in
    #   aiobotocore
    #   taskiq
aioitertools==0.13.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
alembic==1.18.3
    # via
    #   -r requirements.in
    #   alembic-postgresql-enum
alembic-postgresql-enum==1.9.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
    #   taskiq
    #   watchfiles
asgi-correlation-id==4.3.4
    # via -r requirements.in
asyncpg==0.31.0
    # via -r requirements.in
attrs==25.4.0
    # via aiohttp
backports-zstd==1.3.0
    # via aiohttp
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
boto3==1.35.81
    # via aiobotocore
botocore==1.35.81
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
brotli==1.2.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pycares
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
croniter==6.0.0
    # via -r requirements.in
cryptography==46.0.4
    # via
    #   google-auth
    #   python-jose
distro==1.9.0
    # via
    #   google-genai
    #   openai
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via -r requirements.in
fastapi==0.115.14
    # via
    #   -r requirements.in
    #   taskiq-fastapi
filelock==3.20.3
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gitignore-parser==0.1.13
    # via taskiq
google-auth[requests]==2.48.0
    # via google-genai
google-genai==1.62.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   google-genai
    #   openai
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
izulu==0.50.0
    # via taskiq
jiter==0.13.0
    # via openai
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
openai==2.17.0
    # via -r requirements.in
orjson==3.11.7
    # via taskiq
packaging==26.0
    # via
    #   asgi-correlation-id
    #   taskiq
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pillow==12.1.0
    # via
    #   -r requirements.in
    #   pillow-heif
pillow-heif==1.2.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.11
    # via -r requirements.in
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycares==5.0.1
    # via aiodns
pycparser==3.0
    # via cffi
pycron==3.2.0
    # via taskiq
pydantic==2.11.10
    # via
    #   -r requirements.in
    #   fastapi
    #   google-genai
    #   openai
    #   sqlmodel
    #   taskiq
pydantic-core==2.33.2
    # via pydantic
pypdf==6.7.0
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   croniter
python-dotenv==1.2.1
    # via uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.22
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pytz==2025.2
    # via croniter
pyyaml==6.0.3
    # via uvicorn
redis==7.1.0
    # via taskiq-redis
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
sqlalchemy==2.0.46
    # via
    #   -r requirements.in
    #   alembic
    #   alembic-postgresql-enum
    #   sqlmodel
sqlmodel==0.0.32
    # via -r requirements.in
starlette==0.46.2
    # via
    #   asgi-correlation-id
    #   fastapi
stringcase==1.2.0
    # via -r requirements.in
structlog==25.5.0
    # via -r requirements.in
taskiq[orjson,reload]==0.12.1
    # via
    #   -r requirements.in
    #   taskiq-fastapi
    #   taskiq-redis
taskiq-dependencies==1.5.7
    # via
    #   -r requirements.in
    #   taskiq
taskiq-fastapi==0.4.0
    # via -r requirements.in
taskiq-redis==1.2.2
    # via -r requirements.in
tenacity==9.1.4
    # via google-genai
text-unidecode==1.3
    # via python-slugify
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   fastapi
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchdog==6.0.0
    # via taskiq
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   google-genai
    #   uvicorn
wrapt==1.17.3
    # via aiobotocore
yarl==1.22.0
    # via aiohttp
